新闻详情错乱Bug修复

+++ 文明嘉兴/Model/wenmingbobao/News.h
@@ -33,9 +33,9 @@
 
 
 
-+(void)getDataWithPageNum:(int)pageNum;
++(void)getDataWithPageNum:(NSInteger)pageNum;
 
--(void)getDataWithPageNum:(int)pageNum;
+-(void)getDataWithPageNum:(NSInteger)pageNum;
